backend/requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.11.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.8
h11==0.14.0
idna==3.10
mpmath==1.3.0
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.3
typing_extensions==4.12.2
uvicorn==0.34.0
